Extract all MarkdownHover computation to MarkdownHoverParticipant
